fix types

+++ fitness-app/src/server/programs/deleteProgram.ts
@@ -1,13 +1,26 @@
 "use server"
 
 import { prisma } from "@/lib/prisma"
+import { getOrCreateCurrentUser } from "../users/getOrCreateCurrentUser"
 
 export async function deleteProgram(programId : string) {
 
-    await prisma.program.delete({
+    const user = await getOrCreateCurrentUser();
+
+    const program = await prisma.program.findUnique({
+        where:{
+            id: programId
+        }
+    })
+
+    if(program?.userId === user.id){
+        await prisma.program.delete({
         where: {
             id: programId
         }
     })
+    }
+
+    return { success: true }
 
 }
