+++ fitness-app/src/server/programs/deleteTemplate.ts
@@ -1,11 +1,16 @@
 "use server";
 
 import { prisma } from "@/lib/prisma";
+import { getOrCreateCurrentUser } from "../users/getOrCreateCurrentUser";
+
 
 export async function deleteTemplate(templateId: string) {
+
+  const user = await getOrCreateCurrentUser();
+
   const template = await prisma.programTemplate.findUnique({
     where: { id: templateId },
-    select: { sptemplate: true },
+    select: { sptemplate: true, userId: true },
   });
 
   if (!template) {
@@ -14,6 +19,10 @@ export async function deleteTemplate(templateId: string) {
 
   if (template.sptemplate === true) {
     throw new Error("SP templates cannot be deleted.");
+  }
+
+  if(template.userId !== user.id){
+    throw new Error("User does not own this program");
   }
 
   await prisma.programTemplate.delete({
